# services/product-catalog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manages products, SKUs, categories, and images for ZapMarket.",
        "title": "Product Catalog Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/products/{product_id}/images": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-images"
                ],
                "summary": "Get images by product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.ProductImage"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-images"
                ],
                "summary": "Add product image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Image payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateProductImageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ProductImage"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product_id}/images/sku/{sku_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-images"
                ],
                "summary": "Get images by SKU",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "SKU UUID",
                        "name": "sku_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.ProductImage"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product_id}/images/{id}": {
            "delete": {
                "tags": [
                    "product-images"
                ],
                "summary": "Delete product image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product_id}/images/{id}/position": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product-images"
                ],
                "summary": "Update image position",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product UUID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Position payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateImagePositionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ProductImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_zapmarket_zapmarket_services_product-catalog-service_internal_handler_http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/http.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "http.CreateProductImageRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "sku_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "http.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "http.UpdateImagePositionRequest": {
            "type": "object",
            "properties": {
                "position": {
                    "type": "integer"
                }
            }
        },
...
    }
}